refactor(import/notion): split the importer into structure + collection modules
The Notion importer had grown to ~800 lines doing two jobs. Extract the
database-property handling (extractProperties, schema/definition building, date
reconciliation, relation/file application) into collection.ts, the shared
path/id/resource helpers into paths.ts, and the data types into model.ts. importer.ts
keeps page parsing, the note tree and link/image/attachment rewriting, and delegates
property application to collection.ts. Pure move — no behavior change (123 tests).
